refactor: entities/user 리팩토링
