+++ src/entities/user/index.ts
@@ -1,3 +1,4 @@
-export { getUserInfo } from './api/userApi'
-export { type UserInfo } from './api/types'
+export { userApi } from './api/userApi'
+export { type User } from './model/types'
 export { useRefetchUserInfo } from './hooks/useRefetchUserInfo'
+export { useUserInfoStore } from './model/userInfoStore'
