+++ src/entities/user/model/types.ts
@@ -1,4 +1,4 @@
-export type UserInfo = {
+export type User = {
   id: number
   nickName: string
   email: string
